Separate composable-functions from MainFragment into a separate composable package
